added page and style
