+++ library/library.js
@@ -2,52 +2,64 @@ export const library = [
   {
     question: "translate to english: Huhn",
     answer: "chicken",
+    category: "food",
   },
   {
     question: "translate to english: Gurke",
     answer: "cucumber",
+    category: "food",
   },
   {
     question: "translate to english: Auto",
     answer: "car",
+    category: "vehicles",
   },
   {
     question: "translate to english: Fahrrad",
     answer: "bicycle",
+    category: "vehicles",
   },
   {
     question: "Eine Flasche Wasser",
     answer: "a bottle of water",
+    category: "food",
   },
   {
     question:
       "Javascript array method that returns a new array iterator object that contains the key/value pairs for each index in the array.",
     answer: "entries",
+    category: "javascript",
   },
   {
     question:
       "Javascript array method that applies a function against an accumulator and each value of the array (from right-to-left) to reduce it to a single value.",
     answer: "reduceright",
+    category: "javascript",
   },
   {
     question:
       "Javascript array method that changes the contents of an array by removing or replacing existing elements and/or adding new elements in place.",
     answer: "splice",
+    category: "javascript",
   },
   {
     question: "What is a group of flamingos called?",
     answer: "flamboyance",
+    category: "flamingofacts",
   },
   {
     question: "What is the nests of flamingos made of?",
     answer: "mud",
+    category: "flamingofacts",
   },
   {
     question: "How large (in cm) can the largest flamingos in the world grow?",
     answer: "187",
+    category: "flamingofacts",
   },
   {
     question: "how many eggs do flamingos lay in a year?",
     answer: "1",
+    category: "flamingofacts",
   },
 ];
