+++ pages/library/index.js
@@ -1,129 +1,110 @@
-import styled from "styled-components";
-import { useEffect, useState } from "react";
-import { useRef } from "react";
+import Link from "next/link";
 import Image from "next/image";
+import styled from "styled-components";
 
-const StyledForm = styled.form`
-  display: grid;
-  gap: 0.2rem;
-  flex-direction: column;
-  grid-template-columns: 50% 50%;
-  margin-top: 16rem;
+const BookPageStyle = styled.ul`
+  position: relative;
+  font-family: Georgia, "Times New Roman", Times, serif;
   color: black;
-  justify-content: center;
-  align-items: center;
-  font-family: Georgia, "Times New Roman", Times, serif;
-`;
+  list-style: none;
 
-const StyledLabel = styled.label`
-  text-align: right;
-`;
+  padding-left: 0;
+  margin-top: 5rem;
+  margin-bottom: 5rem;
+  margin-left: 3rem;
+  margin-right: 3rem;
 
-const StyledInput = styled.input`
-  background-color: transparent;
-`;
-const StyledButton = styled.button`
-  display: flex;
-  position: absolute;
-  margin-top: 2rem;
-  top: 23rem;
-  text-align: center;
-  justify-self: center;
-  justify-self: center;
-  font-family: Georgia, "Times New Roman", Times, serif;
-  font-weight: bold;
-`;
-const StyledSelect = styled.select`
-  background-color: transparent;
-`;
-const StyledPopup = styled.div`
-  position: absolute;
-  border: 2px solid white;
-  top: 30rem;
-  left: 0;
-  text-align: center;
-  width: 100%;
-`;
-
-const StyledBook = styled(Image)`
-  position: absolute;
-  top: 0;
-  left: 0;
-  object-fit: contain;
-  width: 100%;
-  height: 100%;
-  z-index: -2;
-`;
-
-const BackgroundImage = styled(Image)`
-  position: absolute;
-  top: 0;
-  left: 0;
-  opacity: 0.7;
-  z-index: -3;
-  object-fit: cover;
-`;
-
-export default function AddWisdom({ handleNewWisdomSubmit }) {
-  const inputRef = useRef(null);
-  const [popupActive, setPopupActive] = useState(false);
-
-  function handleSubmit(event) {
-    event.preventDefault();
-
-    const formData = new FormData(event.target);
-    const wisdomData = Object.fromEntries(formData);
-    handleNewWisdomSubmit({ ...wisdomData, answeredRight: 0 });
-    setPopupActive(true);
-    setTimeout(() => {
-      setPopupActive(false);
-    }, 1500);
-    event.target.reset();
-    inputRef.current.focus();
+  li {
+    position: relative;
+    margin-top: 1rem;
+    margin-bottom: 2rem;
   }
-  useEffect(() => {
-    inputRef.current.focus();
-  }, []);
+
+  ::first-letter {
+    font-size: 3rem;
+    font-weight: bold;
+  }
+`;
+const BookPageContainer = styled.div`
+  height: calc(100vh - 8rem);
+  margin-top: 3rem;
+  overflow: scroll;
+`;
+const FeatherLink = styled(Link)`
+  position: fixed;
+`;
+const StyledWisdom = styled.div`
+  position: relative;
+
+  li {
+    margin-top: 2rem;
+    margin-bottom: 2rem;
+  }
+`;
+const BookPageImage = styled.div`
+  position: relative;
+  z-index: -2;
+  height: 100%;
+
+  &::before {
+    content: "";
+    position: fixed;
+    top: 0;
+    right: 0;
+    bottom: 0;
+    left: 0;
+    background-image: url("/assets/bookpage.png");
+    background-repeat: repeat-y;
+    background-size: cover;
+    opacity: 0.7;
+    z-index: -1;
+  }
+`;
+
+export default function ViewBook({ library }) {
+  const categories = library
+    .map((wisdom) => wisdom.category)
+    .filter((value, index, self) => self.indexOf(value) === index);
+  console.log(categories);
   return (
     <>
-      <BackgroundImage
-        src="/assets/desktop.png"
-        alt="desktop"
-        width="400"
-        height="740"
-      ></BackgroundImage>
-      <StyledBook
-        src="/assets/openbook.png"
-        alt="book"
-        width="1080"
-        height="1920"
-      />
-      <StyledForm onSubmit={handleSubmit}>
-        <StyledLabel htmlFor="question">Enter Question:</StyledLabel>
-        <StyledInput
-          ref={inputRef}
-          required
-          name="question"
-          type="text"
-        ></StyledInput>
-        <StyledLabel htmlFor="answer">Enter Answer:</StyledLabel>
-        <StyledInput required name="answer" type="text"></StyledInput>
-        <StyledLabel htmlFor="category">Pick a Category:</StyledLabel>
-        <StyledSelect name="category">
-          <option value="Vehicles">Vehicles</option>
-          <option value="Food">Food</option>
-          <option value="Basics">Basics</option>
-          <option value="Javascript">Javascript</option>
-        </StyledSelect>
-        <StyledLabel htmlFor="benfit">Pick a Benefit</StyledLabel>
-        <StyledSelect name="benefit">
-          <option value="health">Health</option>
-          <option value="armor">Armor</option>
-          <option value="stamina">Stamina</option>
-        </StyledSelect>
-        <StyledButton type="submit">SUBMIT</StyledButton>
-      </StyledForm>
-      {popupActive && <StyledPopup>Wisdom Added</StyledPopup>}
+      <BookPageContainer>
+        <BookPageImage>
+          {categories.map((cat, index) => {
+            return (
+              <BookPageStyle key={index}>
+                <h1>
+                  Chapter {index + 1}: {cat}
+                </h1>
+
+                {library
+                  .filter((wisdom) => wisdom.category === cat)
+                  .map((wisdom, index) => (
+                    <StyledWisdom key={index}>
+                      <li>
+                        <h2>Question:</h2>
+                        <p>{wisdom.question}</p>
+                      </li>
+                      <li>
+                        <h2>Answer:</h2>
+                        <p>{wisdom.answer}</p>
+                      </li>
+                      <p>_____________________________</p>
+                    </StyledWisdom>
+                  ))}
+              </BookPageStyle>
+            );
+          })}
+        </BookPageImage>
+      </BookPageContainer>
+      <FeatherLink href="library/add" passHref legacyBehavior>
+        <Image
+          src="/assets/feather.png"
+          alt="feather"
+          width="200"
+          height="70"
+        ></Image>
+      </FeatherLink>
     </>
   );
 }
